% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaHuber-package.R
\docType{package}
\name{adaHuber-package}
\alias{adaHuber-package}
\title{adaHuber: Adaptive Huber Estimation and Regression}
\description{
Huber-type robust estimation for mean, covariance and (penalized) regression.
}
\references{
Ke, Y., Minsker, S., Ren, Z., Sun, Q. and Zhou, W.-X. (2019). User-friendly covariance estimation for heavy-tailed distributions. Statis. Sci., 34, 454-471.

Pan, X., Sun, Q. and Zhou, W.-X. (2021). Iteratively reweighted l1-penalized robust regression. Electron. J. Stat., 15, 3287-3348.

Sun, Q., Zhou, W.-X. and Fan, J. (2020). Adaptive Huber regression. J. Amer. Stat. Assoc., 115, 254-265.

Wang, L., Zheng, C., Zhou, W. and Zhou, W.-X. (2021). A new principle for tuning-free Huber regression. Stat. Sinica, 31, 2153-2177.
}
\author{
Xiaoou Pan <xip024@ucsd.edu> and Wen-Xin Zhou <wez243@ucsd.edu>
}
