% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{huberCov}
\alias{huberCov}
\title{Tuning-free Huber covariance estimation}
\usage{
huberCov(X, epsilon = 1e-05, iteMax = 500L)
}
\arguments{
\item{X}{An \eqn{n}-dimensional data vector.}

\item{epsilon}{The tolerance level in the iterative estimation procedure, iteration will stop when \eqn{|\mu_new - \mu_old| < \epsilon} or \eqn{|\tau_new - \tau_old| < \epsilon}. The defalut value is 1e-5.}

\item{iteMax}{The maximal number of iteration in the iterative estimation procedure, iteration will stop when this number is reached. The defalut value is 500.}
}
\value{
An adaptive Huber covariance matrix estimator with dimension \eqn{d} by \eqn{d}.
}
\description{
The function calculates adaptive Huber covariance estimator from a data sample, with \eqn{\tau} determined by a tuning-free principle.
}
\details{
The observed data are \eqn{X}, which is an \eqn{n}-dimensional vector whose distribution can be asymmetrix and/or heavy-tailed. The function outputs a robust estimator for the covariance matrix of \eqn{X}.
}
\examples{
n = 400
d = 50
X = matrix(rt(n * d, df = 3), n, d) / sqrt(3)
hubCov = huberCov(X)
}
\references{
Wang, L., Zheng, C., Zhou, W. and Zhou, W.-X. (2018). A New Principle for Tuning-Free Huber Regression. Preprint.

Ke, Y., Minsker, S., Ren, Z., Sun, Q. and Zhou, W.-X. (2019). User-Friendly Covariance Estimation for Heavy-Tailed Distributions. Statis. Sci. To appear.
}
\author{
Xiaoou Pan, Wen-Xin Zhou
}
